+++ src/pages/LoginPage.tsx
@@ -7,8 +7,7 @@ import MainContainer from '../styles/layout';
 import { userInfoAtom, jwtAtom } from '../recoil/atoms';
 import isLoginSelector from '../recoil/seletors';
 import { AlignCenter, ColFlexCenter } from '../styles/shared';
-
-const { VITE_API_URL } = import.meta.env;
+import login from '../apis/apis';
 
 function LoginPage() {
   const location = useLocation();
@@ -25,53 +24,12 @@ function LoginPage() {
     }
   }, []);
 
-  const login = async (email: string, password: string) => {
-    let newError = '';
-    try {
-      const response = await fetch(`${VITE_API_URL}/auth/login`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ email, password }),
-      });
-
-      if (!response.ok) {
-        const data = await response.json();
-        const { statusCode, message, error } = data;
-        if (
-          message.includes('No user found for username') ||
-          message.includes('Invalid password')
-        ) {
-          newError = '아이디 또는 패스워드를 확인하세요.';
-        }
-
-        setDisplayError(newError);
-        throw new Error(
-          `서버에 이상이 있습니다 status: ${statusCode} message: ${message}, error: ${error}`,
-        );
-      }
-
-      const jwt = response.headers.get('authorization');
-
-      setJwt(jwt);
-
-      const responseData = await response.json();
-
-      setUserInfo(responseData);
-      navigate('/');
-    } catch (responseError) {
-      if (responseError instanceof Error) {
-        throw new Error(responseError.message);
-      } else {
-        throw new Error('An unexpected errorjwt occurred.');
-      }
-    }
-  };
-
   const handleLogin = async (email: string, password: string) => {
     try {
-      await login(email, password);
+      const { jwt, userInfo } = await login(email, password, setDisplayError);
+      setJwt(jwt);
+      setUserInfo(userInfo);
+      navigate('/');
     } catch (loginError) {
       console.error(loginError);
     }
